refactor(client): extract auth/mod tests to sibling file
